pylint Similar lines in 2 files
